% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchup.R
\name{switchup}
\alias{switchup}
\title{Switch the first pair of values before and after '=' on the current line}
\usage{
switchup()
}
\description{
Switch the first pair of values before and after '=' on the current line
}
